Add AbstractKafkaBaseIT class to hold common variables and methods used by many Kafka IT classes
