+++ nifi-nar-bundles/nifi-kafka-bundle/nifi-kafka-3-integration/src/test/java/org/apache/nifi/kafka/processors/ConsumeKafkaBaseIT.java
@@ -16,26 +16,13 @@
  */
 package org.apache.nifi.kafka.processors;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.kafka.clients.producer.KafkaProducer;
 import org.apache.kafka.clients.producer.ProducerConfig;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.clients.producer.RecordMetadata;
 import org.apache.kafka.common.header.Header;
 import org.apache.kafka.common.serialization.StringSerializer;
-import org.apache.nifi.json.JsonRecordSetWriter;
-import org.apache.nifi.json.JsonTreeReader;
-import org.apache.nifi.kafka.service.Kafka3ConnectionService;
-import org.apache.nifi.kafka.service.api.KafkaConnectionService;
-import org.apache.nifi.reporting.InitializationException;
-import org.apache.nifi.serialization.RecordReaderFactory;
-import org.apache.nifi.serialization.RecordSetWriterFactory;
-import org.apache.nifi.util.TestRunner;
-import org.junit.jupiter.api.BeforeAll;
-import org.testcontainers.containers.KafkaContainer;
-import org.testcontainers.utility.DockerImageName;
 
-import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -47,55 +34,9 @@ import java.util.stream.Collectors;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public abstract class AbstractConsumeKafkaIT {
+public abstract class ConsumeKafkaBaseIT extends AbstractKafkaBaseIT {
 
-    protected static final String CONNECTION_SERVICE_ID = Kafka3ConnectionService.class.getSimpleName();
-
-    protected static final Duration DURATION_POLL = Duration.ofMillis(1000L);
-
-    protected static final KafkaContainer kafkaContainer;
-
-    private static final String IMAGE_NAME = "confluentinc/cp-kafka:7.3.2";
-
-    static {
-        kafkaContainer = new KafkaContainer(DockerImageName.parse(IMAGE_NAME));
-        kafkaContainer.start();
-    }
-
-    protected static ObjectMapper objectMapper;
-
-    @BeforeAll
-    protected static void beforeAll() {
-        objectMapper = new ObjectMapper();
-    }
-
-    protected void addConnectionService(final TestRunner runner) throws InitializationException {
-        final KafkaConnectionService connectionService = new Kafka3ConnectionService();
-        runner.addControllerService(CONNECTION_SERVICE_ID, connectionService);
-        runner.setProperty(connectionService, Kafka3ConnectionService.BOOTSTRAP_SERVERS, kafkaContainer.getBootstrapServers());
-        runner.setProperty(connectionService, Kafka3ConnectionService.MAX_POLL_RECORDS, "1000");
-        runner.enableControllerService(connectionService);
-    }
-
-    protected String addRecordReaderService(final TestRunner runner) throws InitializationException {
-        final String readerId = "record-reader";
-        final RecordReaderFactory readerService = new JsonTreeReader();
-        runner.addControllerService(readerId, readerService);
-        runner.enableControllerService(readerService);
-        runner.setProperty(readerId, readerId);
-        return readerId;
-    }
-
-    protected String addRecordWriterService(final TestRunner runner) throws InitializationException {
-        final String writerId = "record-writer";
-        final RecordSetWriterFactory writerService = new JsonRecordSetWriter();
-        runner.addControllerService(writerId, writerService);
-        runner.enableControllerService(writerService);
-        runner.setProperty(writerId, writerId);
-        return writerId;
-    }
-
-    protected Properties getProducerProperties() {
+    protected Properties getKafkaProducerProperties() {
         final Properties properties = new Properties();
         properties.setProperty(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, kafkaContainer.getBootstrapServers());
         properties.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class.getName());
@@ -106,7 +47,7 @@ public abstract class AbstractConsumeKafkaIT {
     protected void produceOne(final String topic, final Integer partition,
                               final String key, final String value, final List<Header> headers)
             throws ExecutionException, InterruptedException {
-        try (final KafkaProducer<String, String> producer = new KafkaProducer<>(getProducerProperties())) {
+        try (final KafkaProducer<String, String> producer = new KafkaProducer<>(getKafkaProducerProperties())) {
             final ProducerRecord<String, String> record = new ProducerRecord<>(topic, partition, key, value, headers);
             final Future<RecordMetadata> future = producer.send(record);
             final RecordMetadata metadata = future.get();
@@ -119,7 +60,7 @@ public abstract class AbstractConsumeKafkaIT {
     protected void produce(final String topic, final Collection<ProducerRecord<String, String>> records)
             throws ExecutionException, InterruptedException {
         final Collection<RecordMetadata> metadatas = new ArrayList<>();
-        try (final KafkaProducer<String, String> producer = new KafkaProducer<>(getProducerProperties())) {
+        try (final KafkaProducer<String, String> producer = new KafkaProducer<>(getKafkaProducerProperties())) {
             final Collection<Future<RecordMetadata>> futures = records.stream().map(producer::send).collect(Collectors.toList());
             for (final Future<RecordMetadata> future : futures) {
                 metadatas.add(future.get());
