+++ nifi-nar-bundles/nifi-kafka-bundle/nifi-kafka-3-integration/src/test/java/org/apache/nifi/kafka/processors/ConsumeKafkaKeyAttributeIT.java
@@ -41,7 +41,7 @@ import java.util.stream.Stream;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class ConsumeKafkaKeyAttributeIT extends AbstractConsumeKafkaIT {
+public class ConsumeKafkaKeyAttributeIT extends ConsumeKafkaBaseIT {
 
     private static final String RECORD_KEY = ConsumeKafkaKeyAttributeIT.class.getSimpleName();
     private static final String RECORD_VALUE = ConsumeKafkaKeyAttributeIT.class.getName();
@@ -51,7 +51,7 @@ public class ConsumeKafkaKeyAttributeIT extends AbstractConsumeKafkaIT {
     @BeforeEach
     void setRunner() throws InitializationException {
         runner = TestRunners.newTestRunner(ConsumeKafka.class);
-        addConnectionService(runner);
+        addKafkaConnectionService(runner);
 
         runner.setProperty(ConsumeKafka.CONNECTION_SERVICE, CONNECTION_SERVICE_ID);
         runner.setProperty(ConsumeKafka.AUTO_OFFSET_RESET, AutoOffsetReset.EARLIEST.getValue());
