+++ nifi-nar-bundles/nifi-kafka-bundle/nifi-kafka-3-integration/src/test/java/org/apache/nifi/kafka/processors/PublishKafkaBaseIT.java
@@ -16,89 +16,16 @@
  */
 package org.apache.nifi.kafka.processors;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.common.serialization.StringDeserializer;
-import org.apache.nifi.json.JsonRecordSetWriter;
-import org.apache.nifi.json.JsonTreeReader;
-import org.apache.nifi.kafka.service.Kafka3ConnectionService;
-import org.apache.nifi.kafka.service.api.KafkaConnectionService;
-import org.apache.nifi.reporting.InitializationException;
-import org.apache.nifi.serialization.RecordReaderFactory;
-import org.apache.nifi.serialization.RecordSetWriterFactory;
-import org.apache.nifi.util.TestRunner;
-import org.junit.jupiter.api.BeforeAll;
-import org.testcontainers.containers.KafkaContainer;
-import org.testcontainers.utility.DockerImageName;
 
-import java.time.Duration;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.Properties;
 
-public abstract class PublishKafkaBaseIT {
-    public static final String IMAGE_NAME = "confluentinc/cp-kafka:7.3.2";
-
-    protected static final long TIMESTAMP = System.currentTimeMillis();
-
-    protected static final String SERVICE_ID = Kafka3ConnectionService.class.getSimpleName();
-
-    public static final Duration DURATION_POLL = Duration.ofMillis(1000L);
-
-    protected static final KafkaContainer kafka;
-
-    protected static ObjectMapper objectMapper;
-
-    // https://www.testcontainers.org/test_framework_integration/manual_lifecycle_control/
-    static {
-        kafka = new KafkaContainer(DockerImageName.parse(IMAGE_NAME));
-        kafka.start();
-    }
-
-    @BeforeAll
-    protected static void beforeAll() {
-        objectMapper = new ObjectMapper();
-    }
-
-    protected String addKafkaConnectionService(final TestRunner runner) throws InitializationException {
-        final Map<String, String> connectionServiceProps = new HashMap<>();
-        connectionServiceProps.put(Kafka3ConnectionService.BOOTSTRAP_SERVERS.getName(), kafka.getBootstrapServers());
-        final KafkaConnectionService connectionService = new Kafka3ConnectionService();
-        runner.addControllerService(SERVICE_ID, connectionService, connectionServiceProps);
-        runner.enableControllerService(connectionService);
-        return SERVICE_ID;
-    }
-
-    protected String addRecordReaderService(final TestRunner runner) throws InitializationException {
-        final String readerId = "record-reader";
-        final RecordReaderFactory readerService = new JsonTreeReader();
-        runner.addControllerService(readerId, readerService);
-        runner.enableControllerService(readerService);
-        runner.setProperty(readerId, readerId);
-        return readerId;
-    }
-
-    protected String addRecordWriterService(final TestRunner runner) throws InitializationException {
-        final String writerId = "record-writer";
-        final RecordSetWriterFactory writerService = new JsonRecordSetWriter();
-        runner.addControllerService(writerId, writerService);
-        runner.enableControllerService(writerService);
-        runner.setProperty(writerId, writerId);
-        return writerId;
-    }
-
-    protected String addRecordKeyWriterService(final TestRunner runner) throws InitializationException {
-        final String writerId = "record-key-writer";
-        final RecordSetWriterFactory writerService = new JsonRecordSetWriter();
-        runner.addControllerService(writerId, writerService);
-        runner.enableControllerService(writerService);
-        runner.setProperty(writerId, writerId);
-        return writerId;
-    }
+public abstract class PublishKafkaBaseIT extends AbstractKafkaBaseIT {
 
     protected Properties getKafkaConsumerProperties() {
         final Properties properties = new Properties();
-        properties.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, kafka.getBootstrapServers());
+        properties.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, kafkaContainer.getBootstrapServers());
         properties.put(ConsumerConfig.GROUP_ID_CONFIG, "my-group");
         properties.put(ConsumerConfig.ENABLE_AUTO_COMMIT_CONFIG, Boolean.TRUE.toString());
         properties.put(ConsumerConfig.AUTO_COMMIT_INTERVAL_MS_CONFIG, "1000");
